requirements.txt
aniso8601==9.0.1
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
Flask-Session==0.8.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
msgspec==0.19.0
numpy==2.2.1
opencv-python==4.10.0.84
packaging==24.2
PyJWT==2.10.1
pymongo==4.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
Werkzeug==3.1.3
